+++ pkg/api/v1alpha1/activegate/config.go
@@ -1,0 +1,5 @@
+package activegate
+
+import "github.com/Dynatrace/dynatrace-operator/pkg/util/logger"
+
+var log = logger.Get().WithName("activegate-api")
